Store the cached body path as a String, not a PathBuf.
PathBuf is important when reading paths created by other applications,
because they might not be valid Unicode. However, PathBuf is a wrapper
around OsString, which is different things on different platforms,
which means there's no good, cross-platform way to store it in a SQLite
database.

Since we're going to be generating the filenames we use to store cached
downloads, we can guarantee they'll be valid Unicode, and we can use
the standard "unicode string" type in both Rust and SQLite.