Make sure our types implement Debug, PartialEq and Eq.
Of the traits recommended by the API Guidelines, the only other applicable
one is Clone, and we can't easily implement that because the underlying
database connection object isn't clonable. Creating a new connection object
pointing at the same filesystem path *should* work, but I'd rather not add a
`.unwrap()` just for that.